requirements.txt
fastapi>=0.115.0,<1.0.0
uvicorn[standard]>=0.30.0,<1.0.0
pydantic>=2.7.0,<3.0.0
pydantic-settings>=2.2.1,<3.0.0
python-multipart>=0.0.9,<1.0.0
PyYAML>=6.0.1,<7.0.0
azure-cognitiveservices-speech>=1.41.1,<2.0.0
openai>=1.58.1,<2.0.0
requests>=2.32.0,<3.0.0
